[methods.conf] +version, -everything else.
The only valid formats for a line in methods.conf are now
method_name
or
method_name python_version

python_version can be any py\d+ specified in the daemon config, or just
"py" to run on the same version as the daemon.

The "everything else" that is gone is the dep/link stuff.