[runner] Disable GC in methods.
The idea is that most methods do not actually benefit from the GC, but
they may well be significantly slowed by it.
Additionally, as seen in https://bugs.python.org/issue31558
the GC sometimes causes considerable extra COW after fork.
(If prepare_res is something GC-tracked.)

Once gc.freeze is available we will probably want to call that before
splitting the analysis processes if the method has re-enabled gc.

On an artificial benchmark (d = {ix: [[ix]] for ix in range(10000000)}) we
remove about 30% of run time with the GC off.