Breaking change: No more jobid/dataset confusion.
This changes the API for Dataset.chain and Dataset.iterate_list,
stop_jobid is now stop_ds and jobids is now datasets.

This also changes the meaning of stop_ds (previously stop_jobid).

Previously only the actual jobid was honored. If you had a datasetchain
	jid-1/a jid-1/b jid-2/a jid-2/b
and passed jid-1/a as stop_jobid, you only got the two last datasets,
because the jobid part of jid-1/a matches jid-1/b.
Now you get the last three datasets as you would expect.

This touches a fairly large amount of code, but most of it is just
names changing.