[urd] Accept datetime+integer.
Bare integers sort first, but datetimes without integers sort
before the same datetime with an integer.

Note that 2019-12-10+5 < 2019-12-10T00:00:00 though, because
2019-12-10 < 2019-12-10T00:00:00 despite being "the same" time,
and also 2019-12-10 <= 2019-12+5 even though that is of course
not true in any mathematical sense, it's just what you want.

Also allows .microsecs at the end of times, and " " instead of T.