[dataset_unroundrobin] +trigger_column.
If you do specify trigger_column slice switching is delayed until the
value in trigger_column changes. This is so jobs using new_ds get all
lines with the same trigger_column value in the same slice, useful for
a number of algorithms. (Assuming all occurances of the same value are
together.)