[io] Use pty instead of pipe for capturing output.
I know there was some problem with this that made me initially reject
using ptys, but I can't remember what it was (it wasn't that I couldn't
send binary data over it), and it does work.

Since it's clearly the right thing I'm making this change and hoping for
the best.