[setup] Support static linking of zlib.
This will be needed to build a manylinux-compatible wheel, because the
system zlib might be too old. Normally you can use "auditwheel repair" to
bundle shared libraries you need, but since python is already linked to
zlib the symbols that are in the system zlib (everything except gzbuffer)
will be resolved from there, which doesn't work.

This will also allow using zlib-ng, which otherwise would be stopped by the
same symbol aliasing problem.