[dataset_type] Don't abuse PyTuple_SET_ITEM.
The old usage (changing the first item repeatedly) is probably against the
spirit of the API (even though we have the only reference to the tuple when
we do it). On modern python versions PyObject_CallFunctionObjArgs is just
as fast. (And even if it wasn't this only happens in the slow-path for
uncommon charsets.)