[dataset_type] Don't write the files we won't read.
I should really try to cut down on duplication in this code.
Maybe next time.

This is a similar time save when running the tests with many slices
as the previous commit, but probably still rarely matters in real
projects.

General support for datasets missing slice files for empty slices
might be a good idea too.
Maybe next time?