[urd] Failing to write log is a fatal error.
The likely scenario is that the disk is full. Then you probably get a
truncated last line and everything is very sad. Especially if we then
carry on and append more data to that file (after some space is freed
up).

Now we try to remove the partial data again. That can fail (especially
on COW filesystems), but we should at least try.