[shell/parser] ~ means "back", ^ means "up" for jobs.
For datasets this is the same thing (until the next commit), but for
jobs ~ goes to earlier (current) jobs while ^ follows jobs.previous (or
datasets.previous.job if that is unavailable).

So given that you have just run the current version of the tests:

$ ax grep . 'csvimport~9^/X~11'
contents of M
$

first csvimport~9 resolves to the csvimport subjob from csvimport_zip
that imports the S-file from test_csvimport_zip, then ^ follows
datasets.previous (since jobs.previous isn't set) to the csvimport_zip
job, then we take the X dataset from that, go 11 back and arrive at the
M dataset.

Many other ways of specifying that dataset are of course also possible,
e.g. 'test-1136~~^/O~~' if your workdir is test and it was empty before
running the tests.