[build] don't lose errors
This was originally designed with the assumption that only one client will
call status, but now that both board and build do clearing the error once
it's returned is no good.

So now always return the time of the last error (if one has happened) and
keep the traceback around forever.