[shell/parser] jobs can be specified as :urdlist:[entry]
urdlist can be shortened in the obvious way:
	a -> $USER/a/latest
	a/b -> $USER/a/b
	a/b/c -> a/b/c

entry can be left out and defaults to -1. Anything that looks like a
number will be used a a number, anything else as a name (with standard
joblist resolution rules).

likely specifications are things like ":import/2021-04-06:dataset_type"
or maybe just ":import:".

The second : is required to make dataset specifications easier to
understand (in ":a/b/c" c is always a timestamp, never a dataset name.
":a/b:/c" is required for that).