[shell/grep] always use unicode, regardless of column type
this will be bad for performance on bytes columns, but having separate
logic for bytes and unicode is not a good place to be.

(of course this still has separate encoding logic.)

on python 3 this should be completely transparent, still outputing the
exact bytes that were in a bytes column, but on python 2 I'm not even
going to try. (you get what you deserve for still running python 2.)

matching on bytes that couldn't be decoded as valid utf-8 is possible
by using \udcXX to match \xXX.