[dsutil] number represents small numbers more compactly
now is the time to change this, since the dataset version will also
change in the same release.

all previously valid representations mean the same thing they always did
of course, so reading old datasets is no problem.

[-5, 122] now uses only one byte, [-0x8000, 0x7fff] uses 3 bytes,
[-0x80000000, 0x7fffffff] uses 5 bytes, bigger numbers are unchanged.

this does not usually change the (compressed) size on disk much, but it
does increase both reading and writing speed for small numbers.